Extract execution-runtime behaviour test support
Lift the shared execution-runtime behaviour scaffolding into a colocated
support module so the scenario module stays under the size ceiling.
